add struct.h
